+++ plugins/azure-wakamiti-plugin/src/main/java/module-info.java
@@ -8,5 +8,8 @@ module es.iti.wakamiti.azure {
     requires java.net.http;
     requires json.path;
     requires com.fasterxml.jackson.databind;
+    requires org.apache.commons.lang3;
+    requires org.apache.commons.text;
+    requires org.apache.commons.collections4;
 
 }