+++ plugins/rest-wakamiti-plugin/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 
+## [unreleased]
+
+### Changed
+- Upgrade dependencies: `wakamiti-plugin-starter`
+- Change dependency `plexus-utils` by `commons-io`
+
+
 ## [2.6.0] - 2024-09-09
 
 ### Changed